Fix issue where the first attempt at uploading a file would fail
For some reason the .length property of files would be 0, despite files
containing a non-zero number of files. Furthermore, the length as
reported in the console would be 1, which was really odd.

I'm not 100% why this was happening, but I _believe_ that the reason the
console showed 1 is because we were having some sort of race where we
had length == 0 at the time of the code running and it became 1 later.
Wasn't sure how to fix this but saw something at PluralSight[1] where
they were setting the handler for onChanged, not onClick, and trying
that fixed our issue.

[1]: https://www.pluralsight.com/guides/uploading-files-with-reactjs