Remove vulnerable getTokens endpoint and remove CSRF
What we realized is that /getTokens was inherently vulnerable, since an
attacker could simply call it from another domain to get the tokens and
then issue other endpoint calls. We had CSRF protection on it, but that
makes it impossible to use because it's a chicken and egg problem (why
would we need to give the client the CSRF token if it already has it?).

The reason this was all working before was because csurf by default
ignores certain HTTP methods, namely GET.

Finally, the reason we ended up going nuclear here and remove CSRF was
because of reading this bit:
https://github.com/pillarjs/understanding-csrf#adding-them-to-json-ajax-calls

Now, technically, this blurb is not actually completely correct either
(lol). Or so we think. Our reasoning is that this is true if people
could only ever send requests from a browser that enforces CORS, but
this is not true. People could use older browsers that do not enforce
CORS, or simulate browser requests via things like cURL. Furthermore,
even in the browser, there are supposedly ways to bypass CORS
enforcement, e.g., with Navigator.sendBeacon. _However_, since our
endpoints will require the user to have been logged in, these
alternative methods will not work unless the attacker nabs the users
cookies. They cannot do this without compromising the user's system in
some other way.

When it comes using CSRF through the browser, this won't work because
although the attacker can send a request from another page and re-use
the user's cookies, they cannot inspect the cookies (and our cookies are
HTTP-only, so they are immune to XSS). Therefore, the only vulnerability
we have is old browsers that do not enforce CORS. This is probably
better handled by:
* Not caring since we don't aim to make this app some popular thing. We
  don't even support registration for anyone but us, so as long as we
  don't use an old browser we are fine!
* Not supporting older browsers (though I don't know how to reliably do
  this).