Conditionally set the backend API hostname based on REACT_APP_BACKEND_API_HOSTNAME
This lets us build for production without having to sacrifice usability
during development and testing.

REACT_APP_* prefixed variables are auto-magically embedded into the
script at build time as if they were present in the environment. This is
necessary because react is built into a bunch of _static_ files, not
processes running on the box.